// src/app/pages/transfer/view-details-transfer/view-details-vehicles/shared/view-details-vehicles-enum.ts
export enum ViewDetailsVehiclesEnum {
    mainTitle = 'Vehicles',
    drivers = 'Drivers',
    cars = 'Cars',
    trucks = 'Trucks',
    bus = 'Bus',
    totalVehicles = 'Total Vehicles',
    secondTitle = 'Product List',

    buttonFeatures = 'Setup Vehicle Features',
    selectedTransferIdCookieName = 'selectedTransferId',

    valueZero = 0,
    valueOne = 1,
    car = 'Car',
    truck = 'Truck',
    space = ' ',   
}